dom/power/test/test_power_set_screen_enabled.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test Enabling/Disabling Screen with Power Management API</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
  <script type="application/javascript">

  "use strict";

  function testEnableScreen() {
    try {
      navigator.mozPower.screenEnabled = true;
      ok(navigator.mozPower.screenEnabled === true, "Enabled screen.");
    } catch (e) {
      ok(false, "Unexpected exception trying to enable screen.");
    }
  }

  function testDisableScreen() {
    try {
      navigator.mozPower.screenEnabled = false;
      ok(navigator.mozPower.screenEnabled === false, "Disabled screen.");
    } catch (e) {
      ok(false, "Unexpected exception trying to disable screen.");
    }
  }

  function startTests() {

    // Make sure we don't suspend
    navigator.mozPower.cpuSleepAllowed = false;

    testDisableScreen();
    testEnableScreen();

    SimpleTest.finish();
  }

  SimpleTest.expectAssertions(0, 9);
  SimpleTest.waitForExplicitFinish();
  // Currently only applicable on FxOS
  if (navigator.userAgent.indexOf("Mobile") != -1 &&
      navigator.appVersion.indexOf("Android") == -1) {
    startTests();
  } else {
    ok(true, "mozPower on Firefox OS only.");
    SimpleTest.finish();
  }
  </script>
</pre>
</body>
</html>
